refactor admin pages
