+++ constants/productConstants.ts
@@ -8,4 +8,6 @@ export const SHOP_CONSTANTS = {
     'smartwatch',
     'monitores',
   ],
+  validCapacity: ['64GB', '128GB', '256GB', '512GB', '1TB', '2TB'],
+  validRam: ['4GB', '8GB', '12GB', '16GB', '32GB', '64GB'],
 };
